Prevent using uniform location from an old shaders
The code sending mesh uniforms the bound shader uses the "loc"
atrribute from the uniform without checking if the value stored in
"loc" corresponds to a location from a different shader that
the one being currently bound.

This behavior causes an issue when scenes are rendered several times using different shaders.

Resolves: #10